requirements.txt
PyMuPDF         
pdf2image
opencv-python
numpy
requests
python-dotenv